Move shipping tests to the shipping tests directory
